backend/requirements.txt
django==2.1.5
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
django-cors-headers